# docs/architecture/api/openapi.yaml
openapi: 3.1.2
info:
  title: VideoGame Platform Learning MVP API
  version: 1.1.0
  summary: Browser-facing API for the approved learning MVP vertical slice.
  description: |
    Provider-independent, same-origin contract for release discovery, bounded-catalogue
    search, public game details, BFF session discovery, and the authenticated user's
    ratings. It contains only the eight approved method/path operations.

    Authentication navigation under `/auth` and catalogue synchronization are separate
    BFF and operational boundaries and are intentionally not exposed here. Provider
    image binaries are loaded from an approved CDN reference or replaced by a
    product-owned fallback; this API does not copy, proxy, or redistribute them.
servers:
  - url: /api/v1
    description: Same-origin application deployment
...
security: []
paths:
  /releases:
    get:
      tags: [Releases]
      operationId: listReleases
      summary: Browse recent or upcoming releases
      description: |
        Reads a bounded page from the last valid local catalogue publication. The application derives the
        evaluation date and release window in `Europe/Madrid`. Empty, stale,
        review-required, imprecise-date, TBA, and fallback-cover results are valid
        states. PostgreSQL applies filters, total ordering, count, limit, and offset.
      security: []
      parameters:
        - $ref: '#/components/parameters/ReleaseView'
        - $ref: '#/components/parameters/PlatformIdQuery'
        - $ref: '#/components/parameters/RegionIdQuery'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PublicIfNoneMatch'
      responses:
        '200':
          description: A deterministic page of bounded-catalogue releases, possibly empty.
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
            Cache-Control: { $ref: '#/components/headers/PublicCacheControl' }
            ETag: { $ref: '#/components/headers/PublicEntityTag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReleasePage' }
              examples:
                upcoming:
                  summary: Upcoming release with an imprecise date
                  value:
                    view: upcoming
                    evaluatedOn: '2026-07-31'
                    window: { from: '2026-07-31', to: '2026-12-31' }
                    activeFilters: { platformId: platform_ps5, regionId: region_europe }
                    availableFilters:
                      platforms:
                        - { platformId: platform_ps5, name: PlayStation 5 }
                      regions:
                        - { regionId: region_europe, name: Europe }
                    items:
                      - gameId: game_silksong
                        slug: hollow-knight-silksong
                        canonicalTitle: Hollow Knight Silksong
                        primaryCover:
                          kind: provider
                          url: https://images.igdb.com/igdb/image/upload/t_cover_big/coexample.webp
                          alternativeText: Carátula de Hollow Knight Silksong
                          attribution:
                            label: IGDB
                            sourceUrl: https://www.igdb.com/games/hollow-knight-silksong
                        release:
                          releaseId: release_silksong_ps5_eu
                          gameId: game_silksong
                          platform: { platformId: platform_ps5, name: PlayStation 5 }
                          region: { regionId: region_europe, name: Europe }
                          releaseDate: { precision: quarter, value: 2026-Q4 }
                          status: scheduled
                          provenance:
                            sourceKind: external_provider
                            sourceName: IGDB
                            sourceEntityType: release
                          providerUpdatedAt: '2026-07-25T08:10:00Z'
                          lastSyncedAt: '2026-07-30T05:00:00Z'
                          verificationLevel: provider_only
                          reviewStatus: not_required
                          freshnessStatus: fresh
                    page: { number: 1, size: 20, totalItems: 1, totalPages: 1 }
                staleWithFallback:
                  summary: Last valid stale data remains usable with a fallback cover
                  value:
                    view: recent
                    evaluatedOn: '2026-07-31'
                    window: { from: '2026-01-01', to: '2026-07-31' }
                    activeFilters: { platformId: null, regionId: null }
                    availableFilters:
                      platforms:
                        - { platformId: platform_switch, name: Nintendo Switch }
                      regions:
                        - { regionId: region_europe, name: Europe }
                    items:
                      - gameId: game_example
                        slug: juego-de-ejemplo
                        canonicalTitle: Juego de ejemplo
                        primaryCover:
                          kind: fallback
                          url: /assets/covers/fallback.svg
                          alternativeText: Carátula oficial no disponible
                          attribution: null
                        release:
                          releaseId: release_example_switch_eu
                          gameId: game_example
                          platform: { platformId: platform_switch, name: Nintendo Switch }
                          region: { regionId: region_europe, name: Europe }
                          releaseDate: { precision: month, value: 2026-05 }
                          status: released
                          provenance:
                            sourceKind: external_provider
                            sourceName: IGDB
                            sourceEntityType: release
                          providerUpdatedAt: '2026-05-01T09:00:00Z'
                          lastSyncedAt: '2026-05-02T05:00:00Z'
                          lastVerifiedAt: '2026-05-02T05:10:00Z'
                          verificationLevel: verified
                          reviewStatus: required
                          freshnessStatus: stale
                    page: { number: 1, size: 20, totalItems: 1, totalPages: 1 }
                empty:
                  summary: No release matches the active filters
                  value:
                    view: upcoming
                    evaluatedOn: '2026-07-31'
                    window: { from: '2026-07-31', to: '2026-12-31' }
                    activeFilters: { platformId: platform_ps5, regionId: region_europe }
                    availableFilters:
                      platforms:
                        - { platformId: platform_ps5, name: PlayStation 5 }
                      regions:
                        - { regionId: region_europe, name: Europe }
                    items: []
                    page: { number: 1, size: 20, totalItems: 0, totalPages: 0 }
        '304': { $ref: '#/components/responses/PublicNotModified' }
        '400': { $ref: '#/components/responses/MalformedRequest' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '422': { $ref: '#/components/responses/ReleaseValidationFailed' }
        '429': { $ref: '#/components/responses/RateLimitExceeded' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/CatalogueUnavailable' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-vgp_session
      description: |
        Opaque BFF cookie issued with `Secure`, `HttpOnly`, `SameSite=Lax`, `Path=/`,
        and no `Domain` attribute. OAuth/OIDC tokens remain server-side.
  parameters:
...
    ReleaseView:
      name: view
      in: query
      required: true
      schema: { type: string, enum: [recent, upcoming] }
      example: upcoming
    PlatformIdQuery:
      name: platformId
      in: query
      required: false
      schema: { $ref: '#/components/schemas/PlatformId' }
      example: platform_ps5
    RegionIdQuery:
      name: regionId
      in: query
      required: false
      schema: { $ref: '#/components/schemas/RegionId' }
      example: region_europe
    Page:
      name: page
      in: query
      required: false
      description: One-based page; a page beyond the last returns an empty page.
      schema: { type: integer, minimum: 1, default: 1 }
      example: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      example: 20
...
    PublicIfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: Opaque public representation tag for a conditional read.
      schema: { type: string }
      example: '"catalogue-version-opaque"'
...
  headers:
    XCorrelationId:
      description: Opaque correlation ID, matching Problem Details when present.
      required: true
      schema: { $ref: '#/components/schemas/CorrelationId' }
      example: correlation-opaque
    PublicCacheControl:
      description: Explicit public caching; exact freshness is deployment configuration.
      required: true
      schema: { type: string }
      example: public, max-age=60, stale-while-revalidate=300
    NoStoreCacheControl:
      description: Prevents storage of personal, session, command, and error representations.
      required: true
      schema: { type: string, const: no-store }
      example: no-store
    PublicEntityTag:
      description: Opaque validator for a public representation.
      required: true
      schema: { type: string }
      example: '"catalogue-version-opaque"'
...
    RetryAfter:
      description: Retry delay in seconds when a useful value is known.
      required: false
      schema: { type: integer, minimum: 0 }
      example: 60
  schemas:
    GameId:
      type: string
      minLength: 1
      maxLength: 100
      description: Opaque, provider-independent product identifier.
      examples: [game_example]
    ReleaseId:
      type: string
      minLength: 1
      maxLength: 100
      examples: [release_example_ps5_eu]
    PlatformId:
      type: string
      minLength: 1
      maxLength: 100
      examples: [platform_ps5]
    RegionId:
      type: string
      minLength: 1
      maxLength: 100
      examples: [region_europe]
    CorrelationId:
      type: string
      minLength: 1
      maxLength: 200
      examples: [correlation-opaque]
...
    PageMetadata:
      type: object
      required: [number, size, totalItems, totalPages]
      properties:
        number: { type: integer, minimum: 1 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalItems: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, format: int64, minimum: 0 }
      description: One-based page metadata; `totalPages` is zero for an empty result.
    Platform:
      type: object
      required: [platformId, name]
      properties:
        platformId: { $ref: '#/components/schemas/PlatformId' }
        name: { type: string, minLength: 1 }
    Region:
      type: object
      required: [regionId, name]
      properties:
        regionId: { $ref: '#/components/schemas/RegionId' }
        name: { type: string, minLength: 1 }
    Provenance:
      type: object
      required: [sourceKind, sourceName, sourceEntityType]
      properties:
        sourceKind:
          type: string
          enum: [external_provider, product_curated, official_source]
        sourceName: { type: string, minLength: 1 }
        sourceEntityType: { type: string, minLength: 1 }
      description: Normalized provenance without provider IDs or raw provider records.
    DayReleaseDate:
      type: object
      required: [precision, value]
      properties:
        precision: { type: string, pattern: '^day$' }
        value: { type: string, format: date }
      unevaluatedProperties: false
    MonthReleaseDate:
      type: object
      required: [precision, value]
      properties:
        precision: { type: string, pattern: '^month$' }
        value: { type: string, pattern: '^[0-9]{4}-(0[1-9]|1[0-2])$' }
      unevaluatedProperties: false
    QuarterReleaseDate:
      type: object
      required: [precision, value]
      properties:
        precision: { type: string, pattern: '^quarter$' }
        value: { type: string, pattern: '^[0-9]{4}-Q[1-4]$' }
      unevaluatedProperties: false
    YearReleaseDate:
      type: object
      required: [precision, value]
      properties:
        precision: { type: string, pattern: '^year$' }
        value: { type: string, pattern: '^[0-9]{4}$' }
      unevaluatedProperties: false
    UnknownReleaseDate:
      type: object
      required: [precision, value]
      properties:
        precision: { type: string, pattern: '^unknown$' }
        value: { type: 'null' }
      unevaluatedProperties: false
    ReleaseDate:
      oneOf:
        - $ref: '#/components/schemas/DayReleaseDate'
        - $ref: '#/components/schemas/MonthReleaseDate'
        - $ref: '#/components/schemas/QuarterReleaseDate'
        - $ref: '#/components/schemas/YearReleaseDate'
        - $ref: '#/components/schemas/UnknownReleaseDate'
      description: Closed date/precision union; missing precision is never invented.
    Release:
      type: object
      required: [releaseId, gameId, platform, region, releaseDate, status, provenance, lastSyncedAt, verificationLevel, reviewStatus, freshnessStatus]
      properties:
        releaseId: { $ref: '#/components/schemas/ReleaseId' }
        gameId: { $ref: '#/components/schemas/GameId' }
        platform: { $ref: '#/components/schemas/Platform' }
        region: { $ref: '#/components/schemas/Region' }
        releaseDate: { $ref: '#/components/schemas/ReleaseDate' }
        status:
          type: string
          enum: [announced, scheduled, released, delayed, cancelled, unknown]
        provenance: { $ref: '#/components/schemas/Provenance' }
        providerUpdatedAt: { type: string, format: date-time }
        lastSyncedAt: { type: string, format: date-time }
        lastVerifiedAt: { type: string, format: date-time }
        verificationLevel: { type: string, enum: [provider_only, verified] }
        reviewStatus: { type: string, enum: [not_required, required] }
        freshnessStatus: { type: string, enum: [fresh, stale] }
...
    Attribution:
      type: object
      required: [label, sourceUrl]
      properties:
        label: { type: string, minLength: 1 }
        sourceUrl:
          type: string
          format: uri
          pattern: '^https://www\.igdb\.com/games/'
    ProviderCover:
      type: object
      required: [kind, url, alternativeText, attribution]
      properties:
        kind: { type: string, pattern: '^provider$' }
        url:
          type: string
          format: uri
          pattern: '^https://images\.igdb\.com/igdb/image/upload/'
        alternativeText: { type: string, minLength: 1 }
        attribution: { $ref: '#/components/schemas/Attribution' }
      unevaluatedProperties: false
    FallbackCover:
      type: object
      required: [kind, url, alternativeText, attribution]
      properties:
        kind: { type: string, pattern: '^fallback$' }
        url:
          type: string
          format: uri-reference
          pattern: '^/assets/covers/'
        alternativeText: { type: string, minLength: 1 }
        attribution: { type: 'null' }
      unevaluatedProperties: false
    Cover:
      oneOf:
        - $ref: '#/components/schemas/ProviderCover'
        - $ref: '#/components/schemas/FallbackCover'
...
    ReleaseItem:
      type: object
      required: [gameId, slug, canonicalTitle, primaryCover, release]
      properties:
        gameId: { $ref: '#/components/schemas/GameId' }
        slug: { type: string, minLength: 1 }
        canonicalTitle: { type: string, minLength: 1 }
        primaryCover: { $ref: '#/components/schemas/Cover' }
        release: { $ref: '#/components/schemas/Release' }
    ActiveFilters:
      type: object
      required: [platformId, regionId]
      properties:
        platformId:
          oneOf:
            - $ref: '#/components/schemas/PlatformId'
            - { type: 'null' }
        regionId:
          oneOf:
            - $ref: '#/components/schemas/RegionId'
            - { type: 'null' }
    AvailableFilters:
      type: object
      required: [platforms, regions]
      properties:
        platforms:
          type: array
          items: { $ref: '#/components/schemas/Platform' }
        regions:
          type: array
          items: { $ref: '#/components/schemas/Region' }
    ReleaseWindow:
      type: object
      required: [from, to]
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
    ReleasePage:
      type: object
      required: [view, evaluatedOn, window, activeFilters, availableFilters, items, page]
      properties:
        view: { type: string, enum: [recent, upcoming] }
        evaluatedOn: { type: string, format: date }
        window: { $ref: '#/components/schemas/ReleaseWindow' }
        activeFilters: { $ref: '#/components/schemas/ActiveFilters' }
        availableFilters: { $ref: '#/components/schemas/AvailableFilters' }
        items:
          type: array
          items: { $ref: '#/components/schemas/ReleaseItem' }
        page: { $ref: '#/components/schemas/PageMetadata' }
...
    ProblemCode:
      type: string
      enum:
        - REQUEST_MALFORMED
        - REQUEST_PROPERTY_UNKNOWN
        - REQUEST_PARAMETER_UNKNOWN
        - PAGINATION_INVALID
        - PRECONDITION_REQUIRED
        - CSRF_VALIDATION_FAILED
        - REQUEST_TOO_LARGE
        - MEDIA_TYPE_UNSUPPORTED
        - REPRESENTATION_NOT_ACCEPTABLE
        - METHOD_NOT_ALLOWED
        - RATE_LIMIT_EXCEEDED
        - INTERNAL_ERROR
        - AUTHENTICATION_REQUIRED
        - GAME_NOT_FOUND
        - RATING_NOT_FOUND
        - RATING_VALUE_INVALID
        - RATING_NOT_ELIGIBLE
        - RELEASE_DATA_REVIEW_REQUIRED
        - SEARCH_QUERY_INVALID
        - FILTER_INVALID
        - PLATFORM_NOT_SUPPORTED
        - REGION_NOT_SUPPORTED
        - SORT_INVALID
        - RATING_ALREADY_EXISTS
        - RATING_WRITE_CONFLICT
        - CATALOGUE_NOT_READY
        - CATALOGUE_READ_FAILED
        - RATING_STATISTICS_READ_FAILED
        - PERSONAL_RATINGS_READ_FAILED
        - RATING_WRITE_FAILED
    ErrorCategory:
      type: string
      enum: [validation, authentication, authorization, not_found, business_rule, conflict, dependency, technical]
    Violation:
      type: object
      required: [pointer, code, message]
      properties:
        pointer: { type: string, minLength: 1 }
        code: { $ref: '#/components/schemas/ProblemCode' }
        message:
          type: string
          minLength: 1
          description: Safe mutable guidance; clients do not parse it.
    Problem:
      type: object
      required: [type, title, status, detail, instance, code, category, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string, minLength: 1 }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string, minLength: 1 }
        instance: { type: string, format: uri-reference }
        code: { $ref: '#/components/schemas/ProblemCode' }
        category: { $ref: '#/components/schemas/ErrorCategory' }
        correlationId: { $ref: '#/components/schemas/CorrelationId' }
        eligibilityReason:
          type: string
          enum: [NO_COMMERCIAL_RELEASE, RELEASE_NOT_OCCURRED, RELEASE_CANCELLED, RELEASE_DATE_UNCERTAIN, RELEASE_REVIEW_REQUIRED]
        violations:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Violation' }
      description: RFC 9457 Problem Details with stable product extensions.
  responses:
    PublicNotModified:
      description: The cached public representation remains current; no body is returned.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
        Cache-Control: { $ref: '#/components/headers/PublicCacheControl' }
        ETag: { $ref: '#/components/headers/PublicEntityTag' }
    MalformedRequest:
      description: Request path, query, or JSON syntax cannot be parsed.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
        Cache-Control: { $ref: '#/components/headers/NoStoreCacheControl' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            requestMalformed: { $ref: '#/components/examples/RequestMalformed' }
...
    NotAcceptable:
      description: The client rejects every supported representation.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
        Cache-Control: { $ref: '#/components/headers/NoStoreCacheControl' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            representationNotAcceptable: { $ref: '#/components/examples/RepresentationNotAcceptable' }
...
    ReleaseValidationFailed:
      description: Release filters, pagination, or request parameters are invalid.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
        Cache-Control: { $ref: '#/components/headers/NoStoreCacheControl' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            filterInvalid: { $ref: '#/components/examples/FilterInvalid' }
            platformNotSupported: { $ref: '#/components/examples/PlatformNotSupported' }
            regionNotSupported: { $ref: '#/components/examples/RegionNotSupported' }
            paginationInvalid: { $ref: '#/components/examples/PaginationInvalid' }
            parameterUnknown: { $ref: '#/components/examples/RequestParameterUnknown' }
...
    RateLimitExceeded:
      description: An edge abuse limit rejected the request.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
        Cache-Control: { $ref: '#/components/headers/NoStoreCacheControl' }
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            rateLimitExceeded: { $ref: '#/components/examples/RateLimitExceeded' }
    CatalogueUnavailable:
      description: No local snapshot is ready or local catalogue data cannot be read; no provider fallback call is made.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
        Cache-Control: { $ref: '#/components/headers/NoStoreCacheControl' }
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            catalogueNotReady: { $ref: '#/components/examples/CatalogueNotReady' }
            catalogueReadFailed: { $ref: '#/components/examples/CatalogueReadFailed' }
...
    InternalError:
      description: Unexpected failure without a more specific safe code.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationId' }
        Cache-Control: { $ref: '#/components/headers/NoStoreCacheControl' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            internalError: { $ref: '#/components/examples/InternalError' }
  examples:
    RequestMalformed:
      value:
        type: urn:videogame-platform:problem:request-malformed
        title: Request is malformed
        status: 400
        detail: The request syntax could not be parsed.
        instance: urn:videogame-platform:problem-instance:opaque
        code: REQUEST_MALFORMED
        category: validation
        correlationId: correlation-opaque
...
    RepresentationNotAcceptable:
      value:
        type: urn:videogame-platform:problem:representation-not-acceptable
        title: Representation is not acceptable
        status: 406
        detail: Request application/json or application/problem+json.
        instance: urn:videogame-platform:problem-instance:opaque
        code: REPRESENTATION_NOT_ACCEPTABLE
        category: validation
        correlationId: correlation-opaque
...
    FilterInvalid:
      value:
        type: urn:videogame-platform:problem:filter-invalid
        title: Release filter is invalid
        status: 422
        detail: Correct the release view or filter combination.
        instance: urn:videogame-platform:problem-instance:opaque
        code: FILTER_INVALID
        category: validation
        correlationId: correlation-opaque
    PlatformNotSupported:
      value:
        type: urn:videogame-platform:problem:platform-not-supported
        title: Platform is not supported
        status: 422
        detail: Select a platform from the normalized available filters.
        instance: urn:videogame-platform:problem-instance:opaque
        code: PLATFORM_NOT_SUPPORTED
        category: validation
        correlationId: correlation-opaque
    RegionNotSupported:
      value:
        type: urn:videogame-platform:problem:region-not-supported
        title: Region is not supported
        status: 422
        detail: Select a region from the normalized available filters.
        instance: urn:videogame-platform:problem-instance:opaque
        code: REGION_NOT_SUPPORTED
        category: validation
        correlationId: correlation-opaque
...
    PaginationInvalid:
      value:
        type: urn:videogame-platform:problem:pagination-invalid
        title: Pagination is invalid
        status: 422
        detail: Use page at least 1 and pageSize from 1 through 100.
        instance: urn:videogame-platform:problem-instance:opaque
        code: PAGINATION_INVALID
        category: validation
        correlationId: correlation-opaque
        violations:
          - pointer: /query/pageSize
            code: PAGINATION_INVALID
            message: Must be between 1 and 100.
    RequestParameterUnknown:
      value:
        type: urn:videogame-platform:problem:request-parameter-unknown
        title: Request parameter is unknown
        status: 422
        detail: Remove unsupported query parameters.
        instance: urn:videogame-platform:problem-instance:opaque
        code: REQUEST_PARAMETER_UNKNOWN
        category: validation
        correlationId: correlation-opaque
...
    RateLimitExceeded:
      value:
        type: urn:videogame-platform:problem:rate-limit-exceeded
        title: Rate limit exceeded
        status: 429
        detail: Wait before retrying this request.
        instance: urn:videogame-platform:problem-instance:opaque
        code: RATE_LIMIT_EXCEEDED
        category: technical
        correlationId: correlation-opaque
    CatalogueNotReady:
      value:
        type: urn:videogame-platform:problem:catalogue-not-ready
        title: Catalogue is not ready
        status: 503
        detail: No valid local catalogue snapshot has been published yet.
        instance: urn:videogame-platform:problem-instance:opaque
        code: CATALOGUE_NOT_READY
        category: technical
        correlationId: correlation-opaque
    CatalogueReadFailed:
      value:
        type: urn:videogame-platform:problem:catalogue-read-failed
        title: Catalogue read failed
        status: 503
        detail: Local catalogue data cannot currently be read.
        instance: urn:videogame-platform:problem-instance:opaque
        code: CATALOGUE_READ_FAILED
        category: technical
        correlationId: correlation-opaque
...
    InternalError:
      value:
        type: urn:videogame-platform:problem:internal-error
        title: Internal error
        status: 500
        detail: The request could not be completed.
        instance: urn:videogame-platform:problem-instance:opaque
        code: INTERNAL_ERROR
        category: technical
        correlationId: correlation-opaque